Clean up `wdecl` code.
Primea functions should return IO () (though this is not yet enforced).
They are asynchronous so return values make no sense, and must have side
effects so there is a reason to call them.

The only reason the extra slop existed in the first place was so I could
demo the initial concept in JavaScript. If I want that again, I'll add a
new declaration.